+++ R/flatline_forecaster.R
@@ -17,6 +17,7 @@
 #' @param outcome A scalar character for the column name we wish to predict.
 #' @param args_list A list of dditional arguments as created by the
 #'   [flatline_args_list()] constructor function.
+#' @param .verbose If true, prints the arguments used in the workflow.
 #'
 #' @return A data frame of point (and optionally interval) forecasts at a single
 #'   ahead (unique horizon) for each unique combination of `key_vars`.
@@ -30,7 +31,8 @@
 flatline_forecaster <- function(
     epi_data,
     outcome,
-    args_list = flatline_args_list()) {
+    args_list = flatline_args_list(),
+    .verbose = TRUE) {
   validate_forecaster_inputs(epi_data, outcome, "time_value")
   if (!inherits(args_list, c("flat_fcast", "alist"))) {
     cli_stop("args_list was not created using `flatline_args_list().")
@@ -39,6 +41,10 @@ flatline_forecaster <- function(
   ek <- kill_time_value(keys)
   outcome <- rlang::sym(outcome)
 
+  if (.verbose) {
+    cli::cli_inform("Creating Flatline Forecaster Workflow with the following arguments:")
+    print(args_list)
+  }
 
   r <- epi_recipe(epi_data) %>%
     step_epi_ahead(!!outcome, ahead = args_list$ahead, skip = TRUE) %>%